src/detector.py
"""
detector.py
-----------
Módulo principal de detección de piezas usando YOLOv8.

Responsabilidades:
- Cargar el modelo YOLOv8 entrenado.
- Ejecutar inferencia sobre imágenes estáticas o frames de video.
- Devolver resultados estructurados (bounding boxes, clases, confianzas).
- Opcionalmente dibujar las detecciones sobre la imagen.
"""

import os
from pathlib import Path
from typing import Optional, Union

import cv2
import numpy as np
from dotenv import load_dotenv
from ultralytics import YOLO

load_dotenv()


class PieceDetector:
    """Detector de piezas industriales basado en YOLOv8."""

    def __init__(
        self,
        model_path: str = None,
        confidence: float = None,
        iou_threshold: float = None,
        device: str = None,
    ):
        """
        Inicializa el detector cargando el modelo YOLOv8.

        Args:
            model_path: Ruta al archivo de pesos (.pt). Si es None, usa MODEL_PATH del .env.
            confidence: Umbral de confianza mínima. Si es None, usa MODEL_CONFIDENCE del .env.
            iou_threshold: Umbral IoU para NMS. Si es None, usa MODEL_IOU_THRESHOLD del .env.
            device: Dispositivo de inferencia ('cpu', 'cuda', 'mps').
        """
        self.model_path = model_path or os.getenv("MODEL_PATH", "models/best.pt")
        self.confidence = float(confidence or os.getenv("MODEL_CONFIDENCE", 0.5))
        self.iou_threshold = float(iou_threshold or os.getenv("MODEL_IOU_THRESHOLD", 0.45))
        self.device = device or os.getenv("DEVICE", "cpu")
        self.model: Optional[YOLO] = None
        self._load_model()

    def _load_model(self) -> None:
        """Carga el modelo YOLO desde el path configurado."""
        pass

    def detect(self, source: Union[str, np.ndarray]) -> list[dict]:
        """
        Ejecuta detección sobre una imagen o frame.

        Args:
            source: Ruta a imagen (str) o array NumPy (H, W, C) BGR.

        Returns:
            Lista de diccionarios con claves:
                - 'class_id'   (int)
                - 'class_name' (str)
                - 'confidence' (float)
                - 'bbox'       (list[float]): [x1, y1, x2, y2]
        """
        pass

    def detect_batch(self, sources: list[Union[str, np.ndarray]]) -> list[list[dict]]:
        """
        Ejecuta detección sobre un lote de imágenes.

        Args:
            sources: Lista de rutas o arrays.

        Returns:
            Lista de resultados, uno por imagen.
        """
        pass

    def draw_detections(
        self, image: np.ndarray, detections: list[dict]
    ) -> np.ndarray:
        """
        Dibuja bounding boxes y etiquetas sobre una imagen.

        Args:
            image: Array BGR de la imagen original.
            detections: Resultados de detect().

        Returns:
            Imagen con las detecciones dibujadas.
        """
        pass

    def run_video(
        self,
        source: Union[int, str] = 0,
        show: bool = True,
        save_path: Optional[str] = None,
    ) -> None:
        """
        Procesa un flujo de video frame a frame.

        Args:
            source: Índice de cámara (int) o ruta/URL a video.
            show: Si True, muestra el video con detecciones en ventana.
            save_path: Ruta de salida para guardar el video procesado.
        """
        pass


if __name__ == "__main__":
    detector = PieceDetector()
    print("PieceDetector inicializado correctamente.")
    print(f"  Modelo  : {detector.model_path}")
    print(f"  Conf    : {detector.confidence}")
    print(f"  IoU     : {detector.iou_threshold}")
    print(f"  Device  : {detector.device}")
